requirements.txt
google-generativeai
pydantic
pandas
beautifulsoup4
python-dotenv
python-dateutil
